Implement basic calculator functions
The Calculator class operates based on the values stored in three
variables:
 - firstOperand
 - operator
 - secondOperand

When a numerical digit is pressed, the correct operand can be updated
(i.e. appended with the new digit) depending on whether "operator" has
already been stored. If no "operator" exists, the firstOperand is
updated. If "operator" is defined, then the secondOperand is updated.

The Calculator class will attempt to calculate the result of the
equation described by the three variables firstOperand, operator,
secondOperand under two conditions:
 - the equals button is pressed (provided a full operation is described
   by the three variables)
 - an operator button is pressed AND a full operation is ALREADY
   described by the three variables

In the second case, this allows the "chaining" of multiple operations
without the need for the equals button to be pressed. The result of the
previous operation becomes the firstOperand of the next operation.