Refactor to use runtime property names
When user is inputting operands, if statements were previously used to
check which operand property to update. The process can be simplified by
using the new property "currentInputOperand", which is assigned to a
string containing the property name "firstOperand" or "secondOperand".
Functions can then act on this[this.currentInputOperand] directly, with
no if statements required.